+++ cpp/memory_manager.hpp
@@ -4,11 +4,11 @@ namespace memory {
     class MemoryManager {
     public:
         virtual ~MemoryManager() {}
-        virtual void cleanup_matrix(double** matrix, const int &size) = 0;
+        virtual void cleanup_matrix(long double** matrix, const int &size) = 0;
     };
 
     class MemoryManagerImpl : public MemoryManager {
     public:
-        void cleanup_matrix(double** matrix, const int &size) override;
+        void cleanup_matrix(long double** matrix, const int &size) override;
     };
 }
